pyproject.toml
[project]
name = "tgx"
version = "0.1.0"
description = "Personal Telegram archiver/exporter CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.36.0",
    "qrcode>=7.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
tgx = "tgx.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

